Split backend to smaller files
