+++ server/base.js
@@ -1,0 +1,9 @@
+module.exports = {
+  maindb : null,
+  head : {
+    'Content-Type': 'text/plain',
+    "Access-Control-Allow-Origin": "*",
+    "Access-Control-Request-Method": "POST,GET,PUT,DELETE,OPTIONS",
+    "Access-Control-Allow-Headers": "X-Requested-With,Content-Type"
+  }
+}
